Replace setup.py with everything using pip now
