+++ panda/python/core/MANIFEST.in
@@ -1,2 +1,3 @@
 graft pandare
-global-exclude *.py[cod] __pycache__ *.sw*
+prune pandare/include
+global-exclude *.py[cod] __pycache__ *.sw* .git*
