+++ panda/scripts/install_ubuntu.sh
@@ -147,8 +147,8 @@ pushd build
 progress "PANDA is built and ready to use in panda/build/[arch]-softmmu/panda-system-[arch]."
 
 cd ../panda/python/core
-$SUDO python3 -m pip install -r requirements.txt
-$SUDO python3 setup.py install
+python3 create_panda_datatypes.py --install
+$SUDO python3 -m pip install .
 python3 -c "import pandare; panda = pandare.Panda(generic='i386')" # Make sure it worked
 progress "Pypanda successfully installed"
 popd
